big fixed

+++ sponews/src/main/java/com/sponews/dao/BoardDAO.java
@@ -9,6 +9,6 @@ public interface BoardDAO {
 
 	public List<HashMap<String, Object>> getArticles(@Param("offset") int offset);
 	public int setArticle(HashMap<String, Object> requestMap);
-	public HashMap<String, Object> getArticle(@Param("aid") int aid);
+	public HashMap<String, Object> getArticle(@Param("aid") long aid);
 	public void updateViewCount(@Param("aid") int aid);
 }
